ディレクトリ構成をマイクロサービス化
